Close recovery coherence review gaps
